Fix crash when InfiniteGrid.shadow_color is None.
Commit 1baf349 made shadow_color optional on InfiniteGrid, but the
render and render::shadow modules still expected it to be a required
field. This meant that setting shadow_color: None in an InfiniteGrid
would crash in [prepare_grid_shadows](https://github.com/ForesightMiningSoftwareCorporation/bevy_infinite_grid/blob/38fca70b1/src/render/mod.rs#L312)
when the unwrap was called.

A naive solution was just to introduce a fallible "if let Some(...)" to
the call, which did successfully get that function past the unwrap panic,
but later calls to rendering showed that the wgpu BindGroup was trying
to render without any entities to render.

This commit makes two minor changes to the render module. The first, adds
filtering to the iter map in `extract_grid_shadows` that ensures only
GridFrustumIntersects associated with ExtractedInfiniteGrids with
grid.shadow_color that are Some(...) can be returned. The second change,
adds the fallible "if let Some(...)" to `prepare_grid_shadows` so that
it does not panic in the event it receives a `grid.shadow_color` that is
None.